Refactor: Moved module classes to modules/ root
