+++ modules/__init__.py
@@ -1,0 +1,3 @@
+from modules.crawler import Crawler
+from modules.extractor import Extractor
+from modules.visualization import Visualization
